feat(sdk-coin-avaxp): implement Signer Tx builder
BG-56702
