+++ modules/sdk-coin-avaxp/test/unit/lib/exportP2CTxBuilder.ts
@@ -73,7 +73,7 @@ describe('AvaxP Export P2C Tx Builder', () => {
         .getExportBuilder()
         .threshold(testData.EXPORT_P_2_C.threshold)
         .locktime(testData.EXPORT_P_2_C.locktime)
-        .fromPubKey(testData.EXPORT_P_2_C.pAddresses)
+        .sender(testData.EXPORT_P_2_C.pAddresses)
         .amount(testData.EXPORT_P_2_C.amount)
         .externalChainId(testData.EXPORT_P_2_C.targetChainId)
         .memo(testData.EXPORT_P_2_C.memo)
@@ -95,7 +95,7 @@ describe('AvaxP Export P2C Tx Builder', () => {
         .getExportBuilder()
         .threshold(testData.EXPORT_P_2_C.threshold)
         .locktime(testData.EXPORT_P_2_C.locktime)
-        .fromPubKey(testData.EXPORT_P_2_C.pAddresses)
+        .sender(testData.EXPORT_P_2_C.pAddresses)
         .amount(testData.EXPORT_P_2_C.amount)
         .externalChainId(testData.EXPORT_P_2_C.targetChainId)
         .memo(testData.EXPORT_P_2_C.memo)
@@ -118,7 +118,7 @@ describe('AvaxP Export P2C Tx Builder', () => {
         .getExportBuilder()
         .threshold(testData.EXPORT_P_2_C_WITHOUT_CHANGEOUTPUT.threshold)
         .locktime(testData.EXPORT_P_2_C_WITHOUT_CHANGEOUTPUT.locktime)
-        .fromPubKey(testData.EXPORT_P_2_C_WITHOUT_CHANGEOUTPUT.pAddresses)
+        .sender(testData.EXPORT_P_2_C_WITHOUT_CHANGEOUTPUT.pAddresses)
         .amount(testData.EXPORT_P_2_C_WITHOUT_CHANGEOUTPUT.amount)
         .externalChainId(testData.EXPORT_P_2_C_WITHOUT_CHANGEOUTPUT.targetChainId)
         .memo(testData.EXPORT_P_2_C_WITHOUT_CHANGEOUTPUT.memo)
@@ -140,7 +140,7 @@ describe('AvaxP Export P2C Tx Builder', () => {
         .getExportBuilder()
         .threshold(testData.EXPORT_P_2_C_WITHOUT_CHANGEOUTPUT.threshold)
         .locktime(testData.EXPORT_P_2_C_WITHOUT_CHANGEOUTPUT.locktime)
-        .fromPubKey(testData.EXPORT_P_2_C_WITHOUT_CHANGEOUTPUT.pAddresses)
+        .sender(testData.EXPORT_P_2_C_WITHOUT_CHANGEOUTPUT.pAddresses)
         .amount(testData.EXPORT_P_2_C_WITHOUT_CHANGEOUTPUT.amount)
         .externalChainId(testData.EXPORT_P_2_C_WITHOUT_CHANGEOUTPUT.targetChainId)
         .memo(testData.EXPORT_P_2_C_WITHOUT_CHANGEOUTPUT.memo)
